chore: move tableconfig to new config folder
